paginator: Improve page size handling
Decreasing the page size would result in half of the current page size.
This resulted in confusing behaviour if the current page size was capped
at the number of items. This behaviour will always occur if the page
size is a value other than the expected values, so the (requested) page
size is no longer capped, but always a value from our list.

This list has been expanded to give a few more options between 10 and
50, which I think is the sweet spot for most users.