paginator: Ensure pagenumber is correctly set
The value of pageNumber was set straight from the input. It was checked
and changed in getPages(), but not in getItemsList(), causing the number
of items to appear incorrect in some cases. For example, you would see
  81 to 25 of 25
on a page with 25 items, and page=5&items=20

Having a dedicated function for this makes it explicit and will avoid
trouble if ever someone calls setPageNumber() before setCount() or
setPageItems().