ConfirmSnippetTrait: Catch foreign key violations
Instead of throwing a backtrace, we now show a message that an object
cannot be deleted, with some (technical) information about wich object
is referencing.
This could be improved by adding the table names to the translation
files, and translating them into functional object names.

Parsing the exception message is not ideal, but good enough for now.